% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{constructBipartiteGraph}
\alias{constructBipartiteGraph}
\title{Construct Bipartite Graph of Gene-drug Association Patterns}
\usage{
constructBipartiteGraph(
  mat1,
  mat2,
  n_subsample = 1,
  subsampling_ratio = 1,
  parallel = TRUE
)
}
\arguments{
\item{mat1}{an n x p matrix (e.g. for gene expression)}

\item{mat2}{an n x q matrix (e.g. for drug sensitivity)}

\item{n_subsample}{number of times to perform subsampling to generate B}

\item{subsampling_ratio}{fraction of samples to subsample each time}

\item{parallel}{boolean for whether to parallelize subsampling}
}
\value{
a p x q matrix containing information of bipartite relationship
}
\description{
Constructs matrix B containing information of bipartite relationship between mat1 and mat2 (see paper)
}
