% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{createHclust}
\alias{createHclust}
\title{Create Dendrogram}
\usage{
createHclust(labels, mergeMat, nodeHeights)
}
\arguments{
\item{labels}{vector of group names to label dendrogram}

\item{mergeMat}{matrix of dimension [n - 1, 2] that indicates merge events, where n is number of starting groups (see documentation for "merge" matrix contained in hclust class for details)}

\item{nodeHeights}{vector of dendrogram node heights, in order of merge}
}
\value{
object of class dendrogram
}
\description{
Creates dendrogram object from bipartite hierarchical clustering for plotting
}
