% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{dendroOrder}
\alias{dendroOrder}
\title{Merge Order}
\usage{
dendroOrder(mergeMat)
}
\arguments{
\item{mergeMat}{matrix of dimension [n - 1, 2] that indicates merge events, where n is number of starting groups (see documentation for "merge" matrix contained in hclust class for details)}
}
\value{
vector of leaf (group) indices, in order of merge
}
\description{
Creates vector of order by which leaves are merged given matrix of merge events
}
