% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMergeGroupRows}
\alias{getMergeGroupRows}
\title{Given index from hclust merge matrix, return X, Y row indices
corresponding to groups specified by index}
\usage{
getMergeGroupRows(idx, groups, groupMerges)
}
\arguments{
\item{idx}{index from hclust merge}

\item{groups}{list of starting group membership (input to hierBipartite())}

\item{groupMerges}{list of merged groups thus far (groupMerge[[3]] is a
vector of group names from cluster created at third merge)}
}
\value{
vector of row indices corresponding to X, Y
}
\description{
Given index from hclust merge matrix, return X, Y row indices
corresponding to groups specified by index
}
\keyword{internal}
