% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{getSignificantMergedGroups}
\alias{getSignificantMergedGroups}
\title{Select Significant Results from 'HierBipartite'}
\usage{
getSignificantMergedGroups(results, p = 0.05)
}
\arguments{
\item{results}{list of results from bipartite hierarchical clustering}

\item{p}{p-value to filter merged groups by}
}
\value{
list of results from bipartite hierarchical clustering filtered by p-value
}
\description{
Filters bipartite hierarchical clustering merged groups by p-value
}
