% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{matrixDissimilarity}
\alias{matrixDissimilarity}
\title{Matrix dissimilarity}
\usage{
matrixDissimilarity(B1, B2)
}
\arguments{
\item{B1}{a p x q matrix containing information of bipartite relationship}

\item{B2}{a p x q matrix containing information of bipartite relationship}
}
\value{
nucleus norm-based dissimilarity
}
\description{
Computes nucleus norm-based dissimilarity measure between two matrices
}
