% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newMergedGroup}
\alias{newMergedGroup}
\title{Given group indices for a merge from hclust merge matrix, return group names of merged cluster}
\usage{
newMergedGroup(idx1, idx2, groups, groupMerges)
}
\arguments{
\item{idx1}{group index of first group from hclust merge}

\item{idx2}{group index of second group from hclust merge}

\item{groups}{list of starting group membership (input to hierBipartite())}

\item{groupMerges}{list of merged groups thus far (groupMerge[[3]] is a vector of group names from cluster created at third merge)}
}
\value{
vector of group names after merge of group indicated by idx1 and idx2
}
\description{
Given group indices for a merge from hclust merge matrix, return group names of merged cluster
}
\keyword{internal}
