% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{null_distri}
\alias{null_distri}
\title{Null distribution of dissimilarity measures}
\usage{
null_distri(X1, Y1, X2, Y2, n.perm = 100, parallel = TRUE)
}
\arguments{
\item{X1}{an n x p matrix (e.g. for gene expression) for group 1}

\item{Y1}{an n x q matrix (e.g. for drug sensitivity) for group 1}

\item{X2}{an n x p matrix (e.g. for gene expression) for group 2}

\item{Y2}{an n x q matrix (e.g. for drug sensitivity) for group 2}

\item{n.perm}{number of null dissimilarity measures to generate}

\item{parallel}{boolean for whether to parallelize subsampling}
}
\value{
vector of length n.perm of null dissimilarity measures
}
\description{
Generates null distribution of dissimilarity measures between (X1, Y1) and (X2, Y2)
}
