% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulation Dataset}
\format{
A list with the elements of simulated gene expression, simulated drug sensitivities, and group membership.
\describe{
  \item{X}{n x p gene expression matrix}
  \item{Y}{n x q drug sensitivities matrix}
  \item{groups}{List of starting groups. Each group is represented by a vector of row indices for X, Y.}
}
}
\usage{
simulation
}
\description{
Dataset simulated as described in the paper.
}
\examples{
data(simulation)
X = simulation[["X"]]
Y = simulation[["Y"]]
groups = simulation[["groups"]]
}
\keyword{datasets}
