% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierBipartite.R
\name{whichMerge}
\alias{whichMerge}
\title{Merge groups}
\usage{
whichMerge(dissMat)
}
\arguments{
\item{dissMat}{square matrix of nucleus-norm based dissimilarity measure}
}
\value{
list containing column and row number of entry with minimum dissimilarity
}
\description{
Chooses which entry (row, column) of dissimilarity matrix has minimum dissimilarity
}
